refactor: shared 整理通用工具
