--- packages/shared/src/array.js
@@ -1,6 +1,0 @@
-export const compact = (items) => {
-  if (!Array.isArray(items)) {
-    return [];
-  }
-  return items.filter(Boolean);
-};
