+++ packages/shared/src/index.js
@@ -1,6 +1,11 @@
-export * from "./array.js";
-export * from "./dom.js";
-export * from "./route.js";
-export * from "./storage.js";
-export * from "./style.js";
-export * from "./timing.js";
+export * from "./utils/array.js";
+export * from "./utils/cookie.js";
+export * from "./utils/dom.js";
+export * from "./utils/hash.js";
+export * from "./utils/http.js";
+export * from "./utils/query.js";
+export * from "./utils/route.js";
+export * from "./utils/storage.js";
+export * from "./utils/style.js";
+export * from "./utils/timing.js";
+export * from "./utils/url.js";
