+++ packages/shared/src/utils/style.js
@@ -1,4 +1,14 @@
-export const addStyle = (css, { id = "", target = document.head || document.documentElement } = {}) => {
+/**
+ * Adds or updates a style tag, falling back to GM_addStyle when available.
+ *
+ * @param {string} css CSS text to inject.
+ * @param {object} [options] Injection options.
+ * @returns {HTMLElement | unknown} Injected style node or GM_addStyle result.
+ */
+export const addStyle = (
+  css,
+  { id = "", target = document.head || document.documentElement } = {}
+) => {
   if (id) {
     const existing = document.getElementById(id);
     if (existing) {
