+++ packages/shared/src/utils/timing.js
@@ -1,3 +1,10 @@
+/**
+ * Creates a debounced function with a cancel method.
+ *
+ * @param {Function} callback Function to run after the delay.
+ * @param {number} delay Delay in milliseconds.
+ * @returns {Function & { cancel: () => void }} Debounced function.
+ */
 export const debounce = (callback, delay) => {
   let timer = null;
   const debounced = (...args) => {
